renamed error to exceptions
