+++ testbook/tests/test_client.py
@@ -1,12 +1,10 @@
 import pytest
 
 from testbook import notebook_loader
-from testbook.error import CellTagNotFoundError
+from testbook.exceptions import CellTagNotFoundError
 
 
-@pytest.mark.parametrize(
-    "get_call_index_args, expected_result", [(2, 2), ('hello', 1)]
-)
+@pytest.mark.parametrize("get_call_index_args, expected_result", [(2, 2), ('hello', 1)])
 def test_get_cell_index(get_call_index_args, expected_result, notebook_loader):
     with notebook_loader('testbook/tests/resources/inject.ipynb') as notebook:
         assert notebook._get_cell_index(get_call_index_args) == expected_result
